added async, moved to double

+++ include/stag_ros/stag_nodelet.h
@@ -55,8 +55,8 @@ class StagNodelet : public nodelet::Nodelet {
   void loadParameters(const ros::NodeHandle &nh);
   bool getBundleIndex(const int id, int &bundle_index, int &tag_index);
   bool getTagIndex(const int id, int &tag_index);
-  void solvePnp(const std::vector<cv::Point2f> &img,
-      const std::vector<cv::Point3f> &world,
+  void solvePnp(const std::vector<cv::Point2d> &img,
+      const std::vector<cv::Point3d> &world,
       cv::Mat &output
       );
   // ROS Subcribers
